update 'update customer' to 'update person'
