--- offer-manager-domain/src/main/groovy/life/qbic/business/customers/affiliation/list/ListAffiliationsInput.groovy
@@ -1,18 +1,0 @@
-package life.qbic.business.customers.affiliation.list
-
-/**
- * The input interface for the List Affiliations use case.
- *
- * To start the use case, only one method needs to be called.
- *
- * @author Sven Fillinger
- * @since 1.0.0
- */
-interface ListAffiliationsInput {
-
-  /**
-   * Triggers the List Affiliations use case
-   */
-  void listAffiliations()
-
-}
