--- offer-manager-domain/src/main/groovy/life/qbic/business/customers/update/UpdateCustomer.groovy
@@ -1,62 +1,0 @@
-package life.qbic.business.customers.update
-
-import life.qbic.business.customers.create.CreatePersonDataSource
-import life.qbic.business.logging.Logger
-import life.qbic.business.logging.Logging
-import life.qbic.datamodel.dtos.business.Customer
-
-import life.qbic.business.exceptions.DatabaseQueryException
-
-/**
- * This use case updates an existing customer in the system. New Affiliations of the customer are added to the respective table.
- * If other changes are made to the customer, a new customer is created in the system and the old customer is set to inactive.
- *
- * @since: 1.0.0
- */
-class UpdateCustomer {
-
-  private static final Logging log = Logger.getLogger(UpdateCustomer)
-
-  private CreatePersonDataSource dataSource
-  private UpdateCustomerOutput output
-
-  UpdateCustomer(UpdateCustomerOutput output, CreatePersonDataSource dataSource){
-    this.output = output
-    this.dataSource = dataSource
-  }
-
-  void updateCustomer(int customerId, Customer customer) {
-    Customer existingCustomer = dataSource.getPerson(customerId)
-    boolean customerChanged = hasBasicCustomerDataChanged(existingCustomer, customer)
-    try {
-      if(customerChanged) {
-        dataSource.updatePerson(customerId, customer)
-      } else {
-        dataSource.updatePersonAffiliations(customerId, customer.affiliations)
-      }
-      //this exception catching is important to avoid displaying a wrong failure notification
-      try {
-        output.customerUpdated(customer)
-      } catch (Exception e) {
-        log.error(e.message)
-        log.error(e.stackTrace.join("\n"))
-      }
-    } catch(DatabaseQueryException databaseQueryException){
-      output.failNotification(databaseQueryException.message)
-    } catch(Exception unexpected) {
-      log.error(unexpected.message)
-      log.error(unexpected.stackTrace.join("\n"))
-      output.failNotification("Could not update customer.")
-    }
-  }
-
-  // determines if customer properties other than affiliations have changed
-  private static boolean hasBasicCustomerDataChanged(Customer existingCustomer, Customer newCustomer) {
-    boolean noFundamentalChange = existingCustomer.firstName.equals(newCustomer.firstName)
-    && existingCustomer.lastName.equals(newCustomer.lastName)
-    && existingCustomer.emailAddress.equals(newCustomer.emailAddress)
-    && existingCustomer.title.equals(newCustomer.title)
-    
-    return !noFundamentalChange
-  }
-}
